Fix bug in SaveStage function for multistage builds This change should fix the bug in #294, where kaniko wasn't recognizing that a stage would be used in a later build and so wasn't saving it as a tarball.
Each stage of the Dockerfile has a Name and a BaseName (FROM BaseName as
Name), but if a Name isn't specified then it's set to the same value as
BaseName. Our test cases weren't complete enough to catch this
distinction, which is why this bug occurred.

I added more test cases to the unit tests to make sure this fix works.